api/licenses: add statistics based validation
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 2c064d23-2410-4e86-939f-91c56dae2fae
workspace_id: f8fa0c2a-e6ff-4c03-9c23-bc914a63a8f7

+++ api/pkg/internal/inmemory/inmemory_user_repository.go
@@ -44,6 +44,6 @@ func (f *inMemoryUserRepo) UpdatePassword(u *user.User) error {
 	panic("not implemented")
 }
 
-func (f *inMemoryUserRepo) Count(_ context.Context) (int, error) {
-	return len(f.users), nil
+func (f *inMemoryUserRepo) Count(_ context.Context) (uint64, error) {
+	return uint64(len(f.users)), nil
 }
